Arguments parsing
This marks the first big re-write. The foundation for better argument parsing has now been laid as I switched the script over to a more modular, function-based approach. Arguments are somewhat parsed now, though there is still a distinct hiarchy of order...this will be fixed when I switch to getopts in the future. Also made some bug fixes and updated the readme accordingly.