Reverted back to Dropbox
Recognized an error with iCloud implimentation and reverted to Dropbox for the time being. Automated this a bit more. Dropbox will now be installed and started automatically using Cask. Mackup will also backup configs if one doesn't exist. This comes at the expense of taking up a little more time waiting for Dropbox to sync...this will hopefully be fixed in the future but I'm limited with what status messages I can read from Dropbox at the moment.